Rename module to avoid pylint issue.
Because it had the same name, it believed that it was referring to the
root module, not this particular file, and then yielding an error.
