+++ locales/messages/__init__.py
@@ -6,7 +6,6 @@
 Parent file for the messages module.
 """
 
-from . import gettext, schema
+from . import catalog, gettext, schema
 from .path_loader import load
 from .schema import MessagesSchema
-from .messages import Messages, MessagesData, MessagesTree
